# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "This is the API documentation for the Elonica project.",
        "description": "This is the API documentation for the Elonica project.",
        "contact": {
            "email": "pavlito97@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/password/create": {
            "post": {
                "tags": [
                    "password"
                ],
                "description": "Отправляет письмо на почту для вотановления аккаунта.",
                "operationId": "App\\Http\\Api\\Auth\\PasswordResetController::create",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Письмо дял востановления аккаунта отправлено на почту."
                    },
                    "422": {
                        "description": "Ошибка валидации."
                    }
                }
            }
        },
        "/password/find{token}": {
            "get": {
                "tags": [
                    "password"
                ],
                "description": "Ищет и токен востановления пароля.",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токен найден."
                    },
                    "404": {
                        "description": "Токен не найден."
                    },
                    "422": {
                        "description": "Не корректный токен."
                    }
                }
            }
        },
        "/password/reset": {
            "post": {
                "tags": [
                    "password"
                ],
                "description": "Сохраняет новый пароль.",
                "operationId": "App\\Http\\Api\\Auth\\PasswordResetController::reset",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пароль изменен."
                    },
                    "404": {
                        "description": "Токен или email не найден."
                    },
                    "422": {
                        "description": "Некоррректный пароль или пароль совпадает со старым паролем."
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "description": "Регистрирует новый аккаунт.",
                "operationId": "App\\Http\\Api\\Auth\\RegisterController::register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Игрок успешно зарегистрирован."
                    },
                    "404": {
                        "description": "Игрок не найден."
                    }
                }
            }
        },
        "/users/createCharacter": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Регистрирует новый аккаунт.",
                "operationId": "App\\Http\\Api\\Auth\\RegisterController::createCharacter",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "class",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Персонаж успешно создан."
                    },
                    "422": {
                        "description": "Нельзя создать персонажа."
                    }
                }
            }
        },
        "/friends/{id}": {
            "post": {
                "tags": [
                    "friends"
                ],
                "description": "Добавляет игрока в друзья.",
                "operationId": "App\\Http\\Api\\Controllers\\FriendController::add",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Игрок успешно добавлен в друзья."
                    },
                    "422": {
                        "description": "Нельзя добавить самого себя в друзья."
                    },
                    "404": {
                        "description": "Игрок не найден."
                    }
                }
            },
            "delete": {
                "tags": [
                    "friends"
                ],
                "description": "Удаляет игрока из списка друзей.",
                "operationId": "App\\Http\\Api\\Controllers\\FriendController::deleteFriend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Игрок успешно удален из списка друзей."
                    },
                    "404": {
                        "description": "Игрок не найден."
                    }
                }
            }
        },
        "/friends": {
            "get": {
                "tags": [
                    "friends"
                ],
                "description": "Получить список друзей.",
                "operationId": "App\\Http\\Api\\Controllers\\FriendController::getFriends",
                "responses": {
                    "200": {
                        "description": "Список друзей."
                    },
                    "404": {
                        "description": "Список друзей пуст."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}